feat: dynamic route by route rule config

+++ ship-common/src/main/java/cn/sp/constants/MatchMethodEnum.java
@@ -9,27 +9,27 @@ public enum MatchMethodEnum {
     /**
      * =
      */
-    EQUAL(1, "="),
+    EQUAL((byte)1, "="),
     /**
      * regex
      */
-    REGEX(2, "regex"),
+    REGEX((byte)2, "regex"),
     /**
      * like
      */
-    LIKE(3, "like");
+    LIKE((byte)3, "like");
 
 
-    private Integer code;
+    private Byte code;
 
     private String desc;
 
-    MatchMethodEnum(Integer code, String desc) {
+    MatchMethodEnum(Byte code, String desc) {
         this.code = code;
         this.desc = desc;
     }
 
-    public Integer getCode() {
+    public Byte getCode() {
         return code;
     }
